+++ mzpick-front/src/components/Inputbox/index.tsx
@@ -1,5 +1,4 @@
 import React, { ChangeEvent } from 'react'
-
 import './style.css';
 
 interface Props {
@@ -31,11 +30,11 @@ export default function InputBox({
         <div className="input-box">
             <div className="input-area">
                 <div className="input-text">
-                    <input value={value} type={type} placeholder={placeholder}/>
+                    <input value={value} type={type} placeholder={placeholder} onChange={onChange}/>
                     <div className="shadow"></div>
                 </div>
 
-                {buttonName && <div className={`input-button ${value ? 'active' : 'disable'}`}>{buttonName}</div>}
+                {buttonName && <div className={`input-button ${value ? 'active' : 'disable'}`} onClick={onButtonClick}>{buttonName}</div>}
                 {/* {buttonName && <div className={`input-button ${value ? 'active' : 'active'}`}>{buttonName}</div>} */}
             </div>
