+++ mzpick-front/src/views/Auth/style.css
